storage handler simplified
