--- src/datasource/persistency/entryStorageHandler.js
@@ -1,41 +1,0 @@
-
-class StorageHandler {
-
-    readEntries() {
-        try {
-            const data = localStorage.getItem("journalEntries");
-            return data ? JSON.parse(data) : [];
-        } catch (e) {
-            console.error("Failed to parse journal entries:", e);
-            return [];
-        }
-    }
-
-    writeEntries(entries) {
-        try {
-            localStorage.setItem("journalEntries", JSON.stringify(entries));
-        } catch (e) {
-            console.error("Failed to write journal entries:", e);
-        }
-    }
-
-    readStreak() {
-        try {
-            const data = localStorage.getItem("streak");
-            return data ? JSON.parse(data) : null;
-        } catch (e) {
-            console.error("Failed to parse streak:", e);
-            return null;
-        }
-    }
-
-    writeStreak(value) {
-        try {
-            localStorage.setItem("streak", JSON.stringify(value));
-        } catch (e) {
-            console.error("Failed to write journal entries:", e);
-        }
-    }
-}
-
-export default new StorageHandler();
